mmc: allow the host controller to poll for BKOPS completion
In order to allow the card to perform the required BKOPS and prevent
the need for critical BKOPS, we would like to prevent BKOPS interruption
when possible.
In case the controller calls mmc_suspend_host when runtime suspend is
idle, the BKOPS operation will be interrupted. To prevent this we would
like to prevent the runtime suspend idle until BKOPS is completed.
This patch adds a flag to allow the controller to mark if the polling is
required or not.

Change-Id: Ib290c15a1e9c33dcb01c5fbe8061074049eab985
Signed-off-by: Maya Erez <merez@codeaurora.org>